upgrade web build script

+++ build-web.sh
@@ -1,20 +1,32 @@
-cd ~/grovest/chessvariant
-cargo build
+cd ~/grovest/chess
 
-cd ./web/wasm-chess
-rustwasmc build --no-wasi --target deno --out-dir ../app/api/game/wasm ./
-cd ..
+if [ -z $1 ]; then
+    echo "Supply either 1 or 2"
+    exit 1
+fi
 
-cd ./app
-deno run --allow-read --allow-write build.ts
+if [ $1 = "1" ]; then
+    cargo build
 
-# reload files for an initial build
-rm -rf .aleph
-alephjs dev
+    cd ./web/wasm-chess
+    rustwasmc build --no-wasi --target deno --out-dir ../app/api/game/wasm ./
+    cd ..
 
-## server will eventually terminate via error if browser is open
-## if not, issue SIGTERM and run next lines manually
-# adjust build to work
-deno run --allow-read --allow-write postreload.ts development
+    cd ./app
+    deno run --allow-read --allow-write build.ts
 
-alephjs dev
+    # reload files for an initial build
+    rm -rf .aleph
+    alephjs dev
+elif [ $1 = "2" ]; then
+    cd ./web/app
+
+    ## server will eventually terminate via error if browser is open
+    ## if not, issue SIGTERM and run next lines manually
+    # adjust build to work
+    deno run --allow-read --allow-write postreload.ts development
+
+    alephjs dev
+fi
+
+exit 0
